+++ lista4/Lista04Ex04/src/lista04ex04/ContaPoupança.java
@@ -8,6 +8,6 @@ public class ContaPoupança extends Conta {
     
     @Override
     public void atualizarSaldo(float taxa) {
-        this.saldo += (this.saldo * taxa * 3 )/100;
+        this.saldo += ((float)(this.saldo * taxa * 3 )/100);
     }
 }
